# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /ideas:
    get:
      operationId: listIdeas
      tags: [ideas]
      summary: List innovation lab ideas
      responses:
        "200":
          description: Ideas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/IdeaRecord"
    post:
      operationId: createIdea
      tags: [ideas]
      summary: Submit a new idea
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IdeaInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IdeaRecord"
  /ideas/{id}/status:
    patch:
      operationId: updateIdeaStatus
      tags: [ideas]
      summary: Update an idea status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IdeaStatusChange"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IdeaRecord"
...
  /intake/items:
    get:
      operationId: listIntakeItems
      tags: [intake]
      summary: List external intake items
      responses:
        "200":
          description: Intake items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/IntakeItemRecord"
    post:
      operationId: createIntakeItem
      tags: [intake]
      summary: Capture a new external intake item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IntakeItemInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IntakeItemRecord"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
  /intake/items/{id}:
    patch:
      operationId: updateIntakeItem
      tags: [intake]
      summary: Update an intake item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IntakeItemUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IntakeItemRecord"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
  /memory-candidates:
    get:
      operationId: listMemoryCandidates
      tags: [memory]
      summary: List memory preservation candidates
      responses:
        "200":
          description: Memory candidates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MemoryCandidateRecord"
    post:
      operationId: createMemoryCandidate
      tags: [memory]
      summary: Propose a memory preservation candidate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemoryCandidateInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemoryCandidateRecord"
  /memory-candidates/{id}/status:
    patch:
      operationId: updateMemoryCandidateStatus
      tags: [memory]
      summary: Approve or reject a memory candidate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemoryCandidateStatusChange"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemoryCandidateRecord"
...
components:
  schemas:
...
    ApiMessage:
      type: object
      required: [message]
      properties:
        message:
          type: string
...
    IdeaStatus:
      type: string
      enum:
        - draft-idea
        - related-to-existing
        - needs-research
        - needs-carmen-review
        - needs-rose-review
        - approved-for-build
        - parked
    IdeaApprovalRoute:
      type: string
      enum: [rose, carmen, both, none]
    IdeaRecord:
      type: object
      required: [id, title, description, submittedBy, status, momentum, benefits, risks, dependencies, approvalRoute, createdAt]
      properties:
        id: { type: integer }
        title: { type: string }
        description: { type: string }
        submittedBy: { type: string }
        status: { $ref: "#/components/schemas/IdeaStatus" }
        momentum: { type: integer }
        cluster: { type: ["string", "null"] }
        benefits: { type: array, items: { type: string } }
        risks: { type: array, items: { type: string } }
        dependencies: { type: array, items: { type: string } }
        approvalRoute: { $ref: "#/components/schemas/IdeaApprovalRoute" }
        createdAt: { type: string }
    IdeaInput:
      type: object
      required: [title, description, submittedBy, status, momentum, benefits, risks, dependencies, approvalRoute, createdAt]
      properties:
        title: { type: string, minLength: 1 }
        description: { type: string }
        submittedBy: { type: string, minLength: 1 }
        status: { $ref: "#/components/schemas/IdeaStatus" }
        momentum: { type: integer }
        cluster: { type: ["string", "null"] }
        benefits: { type: array, items: { type: string } }
        risks: { type: array, items: { type: string } }
        dependencies: { type: array, items: { type: string } }
        approvalRoute: { $ref: "#/components/schemas/IdeaApprovalRoute" }
        createdAt: { type: string }
    IdeaStatusChange:
      type: object
      required: [status]
      properties:
        status: { $ref: "#/components/schemas/IdeaStatus" }
...
    IntakeSource:
      type: string
      enum: [zoho_cliq, whatsapp, manual]
    IntakeDetectedType:
      type: string
      enum:
        - idea
        - todo
        - build_request
        - decision_candidate
        - blocker
        - question
        - process_update
        - crm_or_zoho_request
        - automation_request
        - rose_carmen_mind_meld
        - company_brain_update_suggestion
        - sensitive_private_item
        - ignore_or_noise
    IntakeDestination:
      type: string
      enum:
        - mind-meld
        - review-queue
        - command-center-task
        - idea-backlog
        - build-registry
        - requirements-registry
        - automation-registry
        - decision-log
        - company-brain-update
        - no-action
    IntakeSensitivity:
      type: string
      enum:
        - normal
        - private_leadership
        - client_sensitive
        - hr_sensitive
        - financial_sensitive
        - legal_sensitive
        - unclear
    IntakeReviewOwner:
      type: string
      enum: [Rose, Carmen, "Rose and Carmen", "Assigned team member", Unassigned]
    IntakeStatus:
      type: string
      enum: [new, needs_review, routed, approved, rejected, archived]
    IntakeDuplicateRisk:
      type: string
      enum: [none, possible, likely]
    IntakeAuditEntry:
      type: object
      required: [id, timestamp, actor, action]
      properties:
        id: { type: string }
        timestamp: { type: string }
        actor: { type: string }
        action: { type: string }
    IntakeItemRecord:
      type: object
      required:
        - id
        - source
        - sourceChannel
        - senderName
        - senderHandle
        - senderRole
        - receivedAt
        - rawMessage
        - cleanedSummary
        - detectedType
        - suggestedDestination
        - sensitivity
        - reviewOwner
        - status
        - duplicateRisk
        - relatedProjectNames
        - reviewerNotes
        - finalActionTaken
        - nextStep
        - classificationReason
        - auditLog
      properties:
        id: { type: integer }
        source: { $ref: "#/components/schemas/IntakeSource" }
        sourceChannel: { type: string }
        senderName: { type: string }
        senderHandle: { type: string }
        senderRole: { type: ["string", "null"] }
        receivedAt: { type: string }
        rawMessage: { type: string }
        cleanedSummary: { type: string }
        detectedType: { $ref: "#/components/schemas/IntakeDetectedType" }
        suggestedDestination: { $ref: "#/components/schemas/IntakeDestination" }
        sensitivity: { $ref: "#/components/schemas/IntakeSensitivity" }
        reviewOwner: { $ref: "#/components/schemas/IntakeReviewOwner" }
        status: { $ref: "#/components/schemas/IntakeStatus" }
        duplicateRisk: { $ref: "#/components/schemas/IntakeDuplicateRisk" }
        relatedProjectNames: { type: array, items: { type: string } }
        reviewerNotes: { type: string }
        finalActionTaken: { type: ["string", "null"] }
        nextStep: { type: string }
        classificationReason: { type: string }
        auditLog:
          type: array
          items: { $ref: "#/components/schemas/IntakeAuditEntry" }
    IntakeItemInput:
      type: object
      required:
        - source
        - sourceChannel
        - senderName
        - senderHandle
        - rawMessage
        - cleanedSummary
        - detectedType
        - suggestedDestination
        - sensitivity
        - reviewOwner
        - status
        - duplicateRisk
        - relatedProjectNames
        - nextStep
        - classificationReason
        - auditLog
      properties:
        source: { $ref: "#/components/schemas/IntakeSource" }
        sourceChannel: { type: string, minLength: 1 }
        senderName: { type: string, minLength: 1 }
        senderHandle: { type: string, minLength: 1 }
        senderRole: { type: ["string", "null"] }
        receivedAt: { type: string }
        rawMessage: { type: string, minLength: 1 }
        cleanedSummary: { type: string, minLength: 1 }
        detectedType: { $ref: "#/components/schemas/IntakeDetectedType" }
        suggestedDestination: { $ref: "#/components/schemas/IntakeDestination" }
        sensitivity: { $ref: "#/components/schemas/IntakeSensitivity" }
        reviewOwner: { $ref: "#/components/schemas/IntakeReviewOwner" }
        status: { $ref: "#/components/schemas/IntakeStatus" }
        duplicateRisk: { $ref: "#/components/schemas/IntakeDuplicateRisk" }
        relatedProjectNames: { type: array, items: { type: string } }
        reviewerNotes: { type: string }
        nextStep: { type: string }
        classificationReason: { type: string }
        auditLog:
          type: array
          items: { $ref: "#/components/schemas/IntakeAuditEntry" }
    IntakeItemUpdate:
      type: object
      properties:
        status: { $ref: "#/components/schemas/IntakeStatus" }
        reviewOwner: { $ref: "#/components/schemas/IntakeReviewOwner" }
        reviewerNotes: { type: string }
        finalActionTaken: { type: ["string", "null"] }
        auditEntry:
          type: object
          required: [actor, action]
          properties:
            actor: { type: string, minLength: 1 }
            action: { type: string, minLength: 1 }
    MemoryDestination:
      type: string
      enum:
        - private-rose-carmen-memory
        - company-brain-suggestion
        - project-note
        - future-idea
        - decision-candidate
        - knowledge-gap-report
    MemoryCandidateStatus:
      type: string
      enum: [proposed, approved, rejected]
    MemoryCandidateRecord:
      type: object
      required: [id, sourceIntakeId, summary, destination, status, sensitive, createdBy, createdAt]
      properties:
        id: { type: integer }
        sourceIntakeId: { type: ["integer", "null"] }
        summary: { type: string }
        destination: { $ref: "#/components/schemas/MemoryDestination" }
        status: { $ref: "#/components/schemas/MemoryCandidateStatus" }
        sensitive: { type: boolean }
        createdBy: { type: string }
        createdAt: { type: string }
    MemoryCandidateInput:
      type: object
      required: [summary, destination, sensitive, createdBy, createdAt]
      properties:
        sourceIntakeId: { type: ["integer", "null"] }
        summary: { type: string, minLength: 1 }
        destination: { $ref: "#/components/schemas/MemoryDestination" }
        sensitive: { type: boolean }
        createdBy: { type: string, minLength: 1 }
        createdAt: { type: string }
    MemoryCandidateStatusChange:
      type: object
      required: [status]
      properties:
        status: { $ref: "#/components/schemas/MemoryCandidateStatus" }
...